Add option to submit composers instead of artists
For classical music, the composer field is what Last.fm will want to
receive. If the files have correct tags (meaning a composer as a
composer, and a performer as an artist), setting the config option

    composerfolder = "path/holding/classical-music"

will submit composer instead of artist.

E.g. Alfred Brendel playing a Beethoven sonata:

    composer: Ludwig van Beethoven
    artist: Alfred Brendel
    title: Sonata

will submit "Ludwig van Beethoven - Sonata" to Last.fm, if the file in
question is in the folder designated as composerfolder.